requirements.txt
requests==2.22.0
google_auth_oauthlib==0.4.1
google_api_python_client==1.12.3
protobuf==3.13.0
